backend/requirements-training.txt
# Model Training Dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
diffusers>=0.25.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0
datasets>=2.16.0
sentencepiece>=0.1.99

# Audio/Video processing
librosa>=0.10.0
soundfile>=0.12.0
opencv-python>=4.8.0

# Image processing
pillow>=10.0.0
imageio>=2.33.0

# Utilities
loguru>=0.7.0
tqdm>=4.66.0
wandb>=0.16.0  # For training visualization (optional)

# Web search integration
googlesearch-python>=1.2.3
requests>=2.31.0
beautifulsoup4>=4.12.0
